+++ app/src/test/java/org/simple/clinic/teleconsultlog/prescription/medicines/TeleconsultMedicinesUpdateTest.kt
@@ -9,8 +9,6 @@ import com.spotify.mobius.test.UpdateSpec.assertThatNext
 import org.junit.Test
 import org.simple.clinic.TestData
 import org.simple.clinic.drugs.PrescribedDrug
-import org.simple.clinic.drugs.search.DrugFrequency
-import org.simple.clinic.drugs.selection.custom.drugfrequency.country.DrugFrequencyLabel
 import org.simple.clinic.teleconsultlog.medicinefrequency.MedicineFrequency
 import java.time.Duration
 import java.util.UUID
@@ -18,14 +16,6 @@ import java.util.UUID
 class TeleconsultMedicinesUpdateTest {
   private val patientUuid = UUID.fromString("134a6669-7b4c-42dd-a763-e3015b7933ff")
   private val model = TeleconsultMedicinesModel.create(patientUuid = patientUuid)
-
-  private val medicineFrequencyToLabelMap = mapOf(
-      null to DrugFrequencyLabel(label = "None"),
-      MedicineFrequency.OD to DrugFrequencyLabel(label = "OD"),
-      MedicineFrequency.BD to DrugFrequencyLabel(label = "BD"),
-      MedicineFrequency.TDS to DrugFrequencyLabel(label = "TDS"),
-      MedicineFrequency.QDS to DrugFrequencyLabel(label = "QDS")
-  )
 
   private val updateSpec = UpdateSpec(TeleconsultMedicinesUpdate())
 
@@ -67,13 +57,13 @@ class TeleconsultMedicinesUpdateTest {
   }
 
   @Test
-  fun `when drug frequency is clicked, then open drug frequency sheet with medicine frequency to frequency choice items map`() {
+  fun `when drug frequency is clicked, then open drug frequency sheet`() {
     val prescription = TestData.prescription(
         uuid = UUID.fromString("b9c52365-2782-4c03-95ac-1c508a11a510")
     )
 
     updateSpec
-        .given(model.medicineFrequencyToLabelMapLoaded(medicineFrequencyToLabelMap))
+        .given(model)
         .whenEvent(DrugFrequencyClicked(prescription))
         .then(assertThatNext(
             hasNoModel(),
@@ -136,25 +126,6 @@ class TeleconsultMedicinesUpdateTest {
         .then(assertThatNext(
             hasNoModel(),
             hasEffects(OpenEditMedicines(patientUuid)),
-        ))
-  }
-
-  @Test
-  fun `when drug frequency choice items are loaded, then update the model with a map of medicine frequency to frequency choice items`() {
-    val drugFrequencyToLabelMap = mapOf(
-        null to DrugFrequencyLabel(label = "None"),
-        DrugFrequency.OD to DrugFrequencyLabel(label = "OD"),
-        DrugFrequency.BD to DrugFrequencyLabel(label = "BD"),
-        DrugFrequency.TDS to DrugFrequencyLabel(label = "TDS"),
-        DrugFrequency.QDS to DrugFrequencyLabel(label = "QDS")
-    )
-
-    updateSpec
-        .given(model)
-        .whenEvent(DrugFrequencyChoiceItemsLoaded(drugFrequencyToLabelMap))
-        .then(assertThatNext(
-            hasModel(model.medicineFrequencyToLabelMapLoaded(medicineFrequencyToLabelMap)),
-            hasNoEffects()
         ))
   }
 }